requirements.txt
# BrickBox 로컬 PC 학습 환경 의존성
# RTX 2070 SUPER 최적화

# PyTorch (CUDA 지원)
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# YOLO 및 컴퓨터 비전
ultralytics>=8.0.0
opencv-python>=4.8.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0
OpenEXR>=3.0.0  # 깊이 맵 파일 읽기 (EXR 형식)

# 데이터 처리
pandas>=2.0.0
numpy>=1.24.0

# API 및 데이터베이스
supabase>=2.0.0
python-dotenv>=1.0.0

# 유틸리티
tqdm>=4.65.0
onnxruntime>=1.15.0
pyyaml>=6.0
requests>=2.31.0

# Jupyter (선택사항)
jupyter>=1.0.0
ipykernel>=6.25.0

# 개발 도구 (선택사항)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
